Update debug-container.md - Round #1
This one was fairly hard to follow throughout. Tried to cut out some repetition, simplify some statements, and help make things a bit easier to understand. Not sure if I accomplished that, though. It felt kind of like the explanation was going in circles a bit, but I can't really explain specifically why I felt that way.