Update environments.md - Round #1
The third to last paragraph really confused me, so I've reworded some things, but it absolutely needs a look over. There were also a few moments where you would make a point and then re-make it in the same paragraph, and I've changed a little bit of that throughout as well.