Update services-yaml.md - Round #1
Should the title actually be "Services.yaml"? If so, you'll want to change that. I just followed what was on the site for the time being.
I also realized I sometimes get confused with the direction the slash is pointing. Ex. "\" vs "/" so that might be weird occasionally, just in general.
I also reworded a few sections to hopefully improve clarity, so you may need to double check and make sure I'm still saying what you intended.