Update global-bind.md
I'm not sure if the "g" and "b" in "global bind" needed to be capitalized so I left it alone. I also wasn't aways sure if "bind" needed to be in fancy ticks when you were referring directly to a global bind in the code. I opted to use them in a place or two.