Update maker-command.md - Round #1
I'm not entirely sure if I formatted every instance of the word "command" correctly. It pops up a lot but you're not always referring to the same thing in the same way, so you might want to check that. I also reworded things in a few spots to remove some repetition. Hopefully it sounds ok and still means the same thing.