Update persist.md - Round #1
There's a line around the middle of the script where "wherever and whenever" pops up and it took all my self control not to sneak in a Shakira joke. :D I did re-word some things that hopefully still make sense. I also couldn't remember if I need to return the line around 50-ish characters when it's a terminal command for proper text wrapping. I didn't do it for the last terminal command in the script, which was kind of long, since I wasn't 100% sure about it.